examples/openai_client/requirements.txt
openai
python-dotenv

# install humanlayer
humanlayer==0.5.8rc4
# or if you want an editable version
# -e ../../
